get rid of Actor dir
