--- liminal/Actor/ActionLike.ts
@@ -1,4 +1,0 @@
-import type { Action } from "../Action.ts"
-import type { Message } from "../actions/messages.ts"
-
-export type ActionLike = Action | Array<Message>
