+++ liminal/actions/EnableTool.ts
@@ -1,7 +1,6 @@
 import type { StandardSchemaV1 } from "@standard-schema/spec"
-import type { Action } from "../Action.ts"
-import type { ActionLike } from "../Actor/ActionLike.ts"
-import type { Actor } from "../Actor/Actor.ts"
+import type { Action, ActionLike } from "../Action.ts"
+import type { Actor } from "../Actor.ts"
 import type { Spec } from "../Spec.ts"
 import { JSONSchemaMemo } from "../util/JSONSchemaMemo.ts"
 import type { JSONValue } from "../util/JSONValue.ts"
