+++ liminal/actions/Isolate.ts
@@ -1,6 +1,6 @@
-import type { ActionLike } from "../Actor/ActionLike.ts"
-import type { ActorLike } from "../Actor/ActorLike.ts"
-import { reduce } from "../Actor/reduce.ts"
+import type { ActionLike } from "../Action.ts"
+import type { ActorLike } from "../Actor.ts"
+import { reduce } from "../reduceActor.ts"
 import { Scope } from "../Scope.ts"
 import type { ExtractSpec, Spec } from "../Spec.ts"
 import { unwrapDeferred } from "../util/unwrapDeferred.ts"
