+++ liminal/actions/SetMessages.ts
@@ -1,7 +1,6 @@
-import type { Action } from "../Action.ts"
-import type { ActionLike } from "../Actor/ActionLike.ts"
-import type { Actor } from "../Actor/Actor.ts"
-import { reduce } from "../Actor/reduce.ts"
+import type { Action, ActionLike } from "../Action.ts"
+import type { Actor } from "../Actor.ts"
+import { reduce } from "../reduceActor.ts"
 import { Scope } from "../Scope.ts"
 import type { ExtractSpec, Spec } from "../Spec.ts"
 import { isPropertyKey } from "../util/isPropertyKey.ts"
