+++ liminal/exec.ts
@@ -1,8 +1,8 @@
+import type { ActionLike } from "./Action.ts"
 import { ActionEvents } from "./ActionEvents.ts"
 import type { EnteredEvent, ExitedEvent } from "./actions/actions_common.ts"
-import type { ActionLike } from "./Actor/ActionLike.ts"
-import type { ActorLike } from "./Actor/ActorLike.ts"
-import { reduce } from "./Actor/reduce.ts"
+import type { ActorLike } from "./Actor.ts"
+import { reduce } from "./reduceActor.ts"
 import { Scope } from "./Scope.ts"
 import type { ExtractSpec, Spec } from "./Spec.ts"
 import type { FromEntries } from "./util/FromEntries.ts"
