+++ liminal/index.ts
@@ -15,7 +15,7 @@ export type {
   ToolImplementation,
   ToolResult,
 } from "./actions/EnableTool.ts"
-export type { ActorLikeArray, ActorLikeRecord, ActorLikes, Fork, ForkArmResults as ForkResult } from "./actions/Fork.ts"
+export type { Fork, ForkArmResults as ForkResult } from "./actions/Fork.ts"
 export type { GetMessages } from "./actions/GetMessages.ts"
 export type { Infer, InferenceRequestedEvent, InferredEvent } from "./actions/Infer.ts"
 export type { Isolate } from "./actions/Isolate.ts"
@@ -36,13 +36,11 @@ export type {
 export type { EmbeddingModelSetEvent, RunEmbed, SetEmbeddingModel } from "./actions/SetEmbeddingModel.ts"
 export type { LanguageModelSetEvent, RunInfer, SetLanguageModel } from "./actions/SetLanguageModel.ts"
 export type { MessagesSetEvent, SetMessages } from "./actions/SetMessages.ts"
-export * from "./Actor/ActionLike.ts"
-export * from "./Actor/Actor.ts"
-export * from "./Actor/ActorLike.ts"
-export * from "./Actor/reduce.ts"
+export * from "./Actor.ts"
 export { exec, type ExecConfig } from "./exec.ts"
 export * as L from "./L.ts"
 export { LiminalAssertionError } from "./LiminalAssertionError.ts"
+export * from "./reduceActor.ts"
 export * from "./Scope.ts"
 export * from "./Spec.ts"
 export * from "./types/non_factories.ts"
